base64ct: extract and refactor with `Encoding` trait
Refactors common encoding functionality into an `Encoding` trait, and
defines each variant as a ZST which impls (or rather, receives a blanket
impl) of the trait.

This makes trait-based abstractions around various encodings possible,
and also eliminates a considerable amount of duplication.
