+++ base64ct/tests/url.rs
@@ -6,7 +6,7 @@ mod common;
 /// URL-safe Base64 with `=` padding
 mod padded {
     use crate::common::*;
-    use base64ct::url::padded::*;
+    use base64ct::Base64Url;
 
     const TEST_VECTORS: &[TestVector] = &[
         TestVector { raw: b"", b64: "" },
@@ -47,20 +47,23 @@ mod padded {
         },
     ];
 
-    impl_tests!();
+    impl_tests!(Base64Url);
 
     #[test]
     fn reject_trailing_whitespace() {
         let input = "QME/vQVMciqjwvIRc8Bp6kl9NSlrzCRl9vnQQQh716k\n";
         let mut buf = [0u8; 1024];
-        assert_eq!(decode(input, &mut buf), Err(Error::InvalidEncoding));
+        assert_eq!(
+            Base64Url::decode(input, &mut buf),
+            Err(Error::InvalidEncoding)
+        );
     }
 }
 
 /// URL-safe Base64 *without* padding
 mod unpadded {
     use crate::common::*;
-    use base64ct::url::unpadded::*;
+    use base64ct::Base64UrlUnpadded;
 
     const TEST_VECTORS: &[TestVector] = &[
         TestVector { raw: b"", b64: "" },
@@ -101,19 +104,25 @@ mod unpadded {
         },
     ];
 
-    impl_tests!();
+    impl_tests!(Base64UrlUnpadded);
 
     #[test]
     fn reject_trailing_whitespace() {
         let input = "QME/vQVMciqjwvIRc8Bp6kl9NSlrzCRl9vnQQQh716k\n";
         let mut buf = [0u8; 1024];
-        assert_eq!(decode(input, &mut buf), Err(Error::InvalidEncoding));
+        assert_eq!(
+            Base64UrlUnpadded::decode(input, &mut buf),
+            Err(Error::InvalidEncoding)
+        );
     }
 
     #[test]
     fn unpadded_reject_trailing_equals() {
         let input = "QME_vQVMciqjwvIRc8Bp6kl9NSlrzCRl9vnQQQh716k=";
         let mut buf = [0u8; 1024];
-        assert_eq!(decode(input, &mut buf), Err(Error::InvalidEncoding));
+        assert_eq!(
+            Base64UrlUnpadded::decode(input, &mut buf),
+            Err(Error::InvalidEncoding)
+        );
     }
 }
